requirements.txt
Pillow>=10.0.0
playwright>=1.40.0
